Use a 10 minute wide moving average to smooth temp readings
The water temperature smoothly varies since it tries to keep
a uniform temperature but the same can't be said for the air
sensor which is senstive to sun, shade and wind related variations.
Using a 10min window to average temp measurements significantly
reduces the noise. The next step would be to shelter the air temp
sensor in a plastic tube or something to block out the influence
of sun, wind and rain.